Refactor DataSource mocking in tests
Refactor the DataSource mocking to remove code duplication

Refs: LIIK-193
